@@ -6,14 +6,13 @@ package main
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path"
 
-	"github.com/open-policy-agent/golang-opa-wasm/opa"
+	"github.com/hungmol/golang-opa-wasm/opa"
 )
 
-// main demonstrates the loading and executnig of OPA produced wasm
+// main demonstrates the loading and executing of OPA produced wasm
 // policy binary. To execute run 'go run main.go .' in the directory
 // of the main.go.
 func main() {
@@ -26,7 +25,7 @@ func main() {
 
 	// Setup the SDK
 
-	policy, err := ioutil.ReadFile(path.Join(directory, "example-1.wasm"))
+	policy, err := os.ReadFile(path.Join(directory, "example-1.wasm"))
 	if err != nil {
 		fmt.Printf("error: %v\n", err)
 		return
@@ -48,7 +47,20 @@ func main() {
 	}
 
 	ctx := context.Background()
-	result, err := rego.Eval(ctx, &input)
+
+	eps, err := rego.Entrypoints(ctx)
+	if err != nil {
+		fmt.Printf("error: %v\n", err)
+		return
+	}
+
+	entrypointID, ok := eps["example/allow"]
+	if !ok {
+		fmt.Println("error: Unable to find entrypoint 'example/allow'")
+		return
+	}
+
+	result, err := rego.Eval(ctx, opa.EvalOpts{Entrypoint: entrypointID, Input: &input})
 	if err != nil {
 		fmt.Printf("error: %v\n", err)
 		return
@@ -56,42 +68,39 @@ func main() {
 
 	fmt.Printf("Policy 1 result: %v\n", result)
 
-	resultBool, err := opa.EvalBool(ctx, rego, &input)
+	// Update the policy on the fly.
+
+	policy, err = os.ReadFile(path.Join(directory, "example-2.wasm"))
 	if err != nil {
 		fmt.Printf("error: %v\n", err)
 		return
 	}
 
-	fmt.Printf("Policy 1 boolean result: %v\n", resultBool)
-
-	// Update the policy on the fly.
-
-	policy, err = ioutil.ReadFile(path.Join(directory, "example-2.wasm"))
+	// Get an updated entrypoint ID, they may have changed!
+	eps, err = rego.Entrypoints(ctx)
 	if err != nil {
 		fmt.Printf("error: %v\n", err)
+		return
+	}
+
+	entrypointID, ok = eps["example/allow"]
+	if !ok {
+		fmt.Println("error: Unable to find entrypoint 'example/allow'")
 		return
 	}
 
 	// Evaluate the new policy.
 
-	if err := rego.SetPolicy(policy); err != nil {
+	if err := rego.SetPolicy(ctx, policy); err != nil {
 		fmt.Printf("error: %v\n", err)
 		return
 	}
 
-	result, err = rego.Eval(ctx, &input)
+	result, err = rego.Eval(ctx, opa.EvalOpts{Entrypoint: entrypointID, Input: &input})
 	if err != nil {
 		fmt.Printf("error: %v\n", err)
 		return
 	}
 
 	fmt.Printf("Policy 2 result: %v\n", result)
-
-	resultBool, err = opa.EvalBool(ctx, rego, &input)
-	if err != nil {
-		fmt.Printf("error: %v\n", err)
-		return
-	}
-
-	fmt.Printf("Policy 2 boolean result: %v\n", resultBool)
 }
