@@ -1,13 +1,16 @@
 // Copyright 2020 The OPA Authors.  All rights reserved.
 // Use of this source code is governed by an Apache2
 // license that can be found in the LICENSE file.
-package http
+
+//go:build opa_wasm
+// +build opa_wasm
+
+package file
 
 import (
 	"bytes"
 	"context"
-	"net/http"
-	"net/http/httptest"
+	"os"
 	"reflect"
 	"sync"
 	"testing"
@@ -17,51 +20,37 @@ import (
 )
 
 func TestFileLoader(t *testing.T) {
-	// Start loader, without having the HTTP content in place.
+	// Assign a temp file.
+
+	f, err := os.CreateTemp("", "test-file-loader")
+	if err != nil {
+		panic(err)
+	}
+
+	defer os.Remove(f.Name())
+
+	// Start loader, without having a file in place.
 
 	var pd testPolicyData
-	loader, err := new(&pd).WithURL("http://localhost:0").WithInterval(10*time.Millisecond, 20*time.Millisecond).Init()
+	loader, err := new(&pd).WithFile(f.Name()).WithInterval(10 * time.Millisecond).Init()
 	if err != nil {
 		t.Fatal(err.Error())
 	}
 
-	ctx, cancel := context.WithCancel(context.Background())
-	go func() {
-		time.Sleep(10 * time.Millisecond)
-		cancel()
-	}()
-
-	if err := loader.Start(ctx); err != context.Canceled {
-		t.Fatalf("missing file not resulting in a correct error: %v", err)
+	ctx := context.Background()
+	if err := loader.Start(ctx); err == nil {
+		t.Fatal("missing file not resulting in an error")
 	}
 
-	// Start again, with the HTTP content in place.
-
-	var mutex sync.Mutex
 	policy := "wasm-policy"
 	var data interface{} = map[string]interface{}{
 		"foo": "bar",
 	}
 
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		mutex.Lock()
-		defer mutex.Unlock()
+	// Start loader, with the file in place.
 
-		if err := bundle.Write(w, bundle.Bundle{
-			Data: data.(map[string]interface{}),
-			Wasm: []byte(policy),
-		}); err != nil {
-			panic(err)
-		}
-	}))
-	defer ts.Close()
+	writeBundle(f.Name(), policy, data)
 
-	loader, err = new(&pd).WithURL(ts.URL).WithInterval(10*time.Millisecond, 20*time.Millisecond).Init()
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-
-	ctx = context.Background()
 	if err := loader.Start(ctx); err != nil {
 		t.Fatalf("unable to start loader: %v", err)
 	}
@@ -70,12 +59,12 @@ func TestFileLoader(t *testing.T) {
 
 	// Reload with updated contents.
 
-	mutex.Lock()
 	policy = "wasm-policy-modified"
 	data = map[string]interface{}{
 		"bar": "foo",
 	}
-	mutex.Unlock()
+
+	writeBundle(f.Name(), policy, data)
 
 	pd.WaitUpdate()
 	pd.CheckEqual(t, policy, &data)
@@ -90,7 +79,7 @@ type testPolicyData struct {
 	updated chan struct{}
 }
 
-func (pd *testPolicyData) SetPolicyData(policy []byte, data *interface{}) error {
+func (pd *testPolicyData) SetPolicyData(_ context.Context, policy []byte, data *interface{}) error {
 	pd.Lock()
 	defer pd.Unlock()
 
@@ -122,4 +111,20 @@ func (pd *testPolicyData) WaitUpdate() {
 	pd.Lock()
 	pd.updated = nil
 	pd.Unlock()
+}
+
+func writeBundle(name string, policy string, data interface{}) {
+	b := bundle.Bundle{
+		Data: data.(map[string]interface{}),
+		Wasm: []byte(policy),
+	}
+
+	var buf bytes.Buffer
+	if err := bundle.Write(&buf, b); err != nil {
+		panic(err)
+	}
+
+	if err := os.WriteFile(name, buf.Bytes(), 0644); err != nil {
+		panic(err)
+	}
 }
