@@ -5,11 +5,11 @@
 package http
 
 import (
-	"fmt"
 	"net/http"
 	"time"
 
-	"github.com/open-policy-agent/golang-opa-wasm/opa"
+	// "github.com/open-policy-agent/opa/internal/wasm/sdk/opa/errors"
+	"github.com/hungmol/golang-opa-wasm/opa/errors"
 )
 
 // WithURL configures the URL to download the bundle from.
@@ -22,7 +22,7 @@ func (l *Loader) WithURL(url string) *Loader {
 // http.DefaultClient is used.
 func (l *Loader) WithClient(client *http.Client) *Loader {
 	if client == nil {
-		l.configErr = fmt.Errorf("client: %w", opa.ErrInvalidConfig)
+		l.configErr = errors.New(errors.InvalidConfigErr, "client")
 		return l
 	}
 
@@ -33,7 +33,7 @@ func (l *Loader) WithClient(client *http.Client) *Loader {
 // WithInterval configures the minimum and maximum delay between bundle downloads.
 func (l *Loader) WithInterval(min, max time.Duration) *Loader {
 	if min > max {
-		l.configErr = fmt.Errorf("interval: %w", opa.ErrInvalidConfig)
+		l.configErr = errors.New(errors.InvalidConfigErr, "interval min > max")
 		return l
 	}
 
@@ -46,7 +46,7 @@ func (l *Loader) WithInterval(min, max time.Duration) *Loader {
 // HTTP request is not modified after the handle invocation.
 func (l *Loader) WithPrepareRequest(prepare func(*http.Request) error) *Loader {
 	if prepare == nil {
-		l.configErr = fmt.Errorf("prepare request: %w", opa.ErrInvalidConfig)
+		l.configErr = errors.New(errors.InvalidConfigErr, "missing prepare")
 		return l
 	}
 
@@ -57,7 +57,7 @@ func (l *Loader) WithPrepareRequest(prepare func(*http.Request) error) *Loader {
 // WithErrorLogger configures an error logger invoked with all the errors.
 func (l *Loader) WithErrorLogger(logger func(error)) *Loader {
 	if logger == nil {
-		l.configErr = fmt.Errorf("logger: %w", opa.ErrInvalidConfig)
+		l.configErr = errors.New(errors.InvalidConfigErr, "missing logger")
 		return l
 	}
 
