@@ -2,13 +2,15 @@
 // Use of this source code is governed by an Apache2
 // license that can be found in the LICENSE file.
 
-package file
+// +build opa_wasm
+
+package http
 
 import (
 	"bytes"
 	"context"
-	"io/ioutil"
-	"os"
+	"net/http"
+	"net/http/httptest"
 	"reflect"
 	"sync"
 	"testing"
@@ -17,38 +19,52 @@ import (
 	"github.com/open-policy-agent/opa/bundle"
 )
 
-func TestFileLoader(t *testing.T) {
-	// Assign a temp file.
-
-	f, err := ioutil.TempFile("", "test-file-loader")
-	if err != nil {
-		panic(err)
-	}
-
-	defer os.Remove(f.Name())
-
-	// Start loader, without having a file in place.
+func TestHTTPLoader(t *testing.T) {
+	// Start loader, without having the HTTP content in place.
 
 	var pd testPolicyData
-	loader, err := new(&pd).WithFile(f.Name()).WithInterval(10 * time.Millisecond).Init()
+	loader, err := newLoader(&pd).WithURL("http://localhost:0").WithInterval(10*time.Millisecond, 20*time.Millisecond).Init()
 	if err != nil {
 		t.Fatal(err.Error())
 	}
 
-	ctx := context.Background()
-	if err := loader.Start(ctx); err == nil {
-		t.Fatal("missing file not resulting in an error")
+	ctx, cancel := context.WithCancel(context.Background())
+	go func() {
+		time.Sleep(10 * time.Millisecond)
+		cancel()
+	}()
+
+	if err := loader.Start(ctx); err != context.Canceled {
+		t.Fatalf("missing file not resulting in a correct error: %v", err)
 	}
 
+	// Start again, with the HTTP content in place.
+
+	var mutex sync.Mutex
 	policy := "wasm-policy"
 	var data interface{} = map[string]interface{}{
 		"foo": "bar",
 	}
 
-	// Start loader, with the file in place.
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		mutex.Lock()
+		defer mutex.Unlock()
 
-	writeBundle(f.Name(), policy, data)
+		if err := bundle.Write(w, bundle.Bundle{
+			Data: data.(map[string]interface{}),
+			Wasm: []byte(policy),
+		}); err != nil {
+			panic(err)
+		}
+	}))
+	defer ts.Close()
 
+	loader, err = newLoader(&pd).WithURL(ts.URL).WithInterval(10*time.Millisecond, 20*time.Millisecond).Init()
+	if err != nil {
+		t.Fatal(err.Error())
+	}
+
+	ctx = context.Background()
 	if err := loader.Start(ctx); err != nil {
 		t.Fatalf("unable to start loader: %v", err)
 	}
@@ -57,12 +73,12 @@ func TestFileLoader(t *testing.T) {
 
 	// Reload with updated contents.
 
+	mutex.Lock()
 	policy = "wasm-policy-modified"
 	data = map[string]interface{}{
 		"bar": "foo",
 	}
-
-	writeBundle(f.Name(), policy, data)
+	mutex.Unlock()
 
 	pd.WaitUpdate()
 	pd.CheckEqual(t, policy, &data)
@@ -77,7 +93,7 @@ type testPolicyData struct {
 	updated chan struct{}
 }
 
-func (pd *testPolicyData) SetPolicyData(policy []byte, data *interface{}) error {
+func (pd *testPolicyData) SetPolicyData(_ context.Context, policy []byte, data *interface{}) error {
 	pd.Lock()
 	defer pd.Unlock()
 
@@ -109,20 +125,4 @@ func (pd *testPolicyData) WaitUpdate() {
 	pd.Lock()
 	pd.updated = nil
 	pd.Unlock()
-}
-
-func writeBundle(name string, policy string, data interface{}) {
-	b := bundle.Bundle{
-		Data: data.(map[string]interface{}),
-		Wasm: []byte(policy),
-	}
-
-	var buf bytes.Buffer
-	if err := bundle.Write(&buf, b); err != nil {
-		panic(err)
-	}
-
-	if err := ioutil.WriteFile(name, buf.Bytes(), 0644); err != nil {
-		panic(err)
-	}
 }
